+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Hide edit and delete buttons for non Tier 1 users
 - Placeholder page for editing decision data
 - Placeholder page for adding decision data
+- Add Regulator column to decision data table for central admin users
 
 ## [release-016] - 2022-04-11